Alta de producto y consultar producto funcionando

+++ include/Producto.h
@@ -13,11 +13,11 @@ using namespace std;
 class Producto {
 public:
     Producto();
-    Producto(string nombre, string descripcion, int id, int stock, int precio, Categoria categoria, Vendedor* vendedor);
+    Producto(string nombre, string descripcion, int id, int stock, float precio, Categoria categoria, Vendedor* vendedor);
     string getNombre();
     string getDescripcion();
     Categoria getCategoria();
-    int getPrecio();
+    float getPrecio();
     int getId();
     int getStock();
     Vendedor* getVendedor() const;
@@ -32,7 +32,7 @@ private:
     string nombre;
     int id;
     int stock;
-    int precio;
+    float precio;
     Categoria categoria;
     Vendedor* vendedor;
 };
